fix variable names, fix comparison between signed and unsigned values
Commit-type: bug-fix, portability-fix
Reported-by: nightly tests

+++ src/sys/classes/viewer/impls/ascii/filev.c
@@ -987,8 +987,8 @@ PetscErrorCode  PetscViewerASCIISynchronizedPrintf(PetscViewer viewer,const char
     va_start(Argp,format);
     ierr = PetscVSNPrintf(string,next->size-2*vascii->tab,format,&fullLength,Argp);CHKERRQ(ierr);
     va_end(Argp);
-    if (fullLength > next->size-2*vascii->tab) {
-      ierr = PetscFree(next->string);CHKERRQ(ierr);
+    if (fullLength > (size_t) (next->size-2*vascii->tab)) {
+      ierr       = PetscFree(next->string);CHKERRQ(ierr);
       next->size = fullLength + 2*vascii->tab;
       ierr       = PetscMalloc1(next->size, &next->string);CHKERRQ(ierr);
       ierr       = PetscMemzero(next->string,next->size);CHKERRQ(ierr);
